+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v3.1.7 (JAN 07, 2022)
+- Applied type narrowing to `BaseMessage`, `BaseChannel` and  `messageCallback`.
+- Replaced all `File` type with the new `FileType` for React Native support.
+- Fixed create channel with cover image error in React Native.
+- Fixed child message's `parentMessage` update on parent message update when local cache is enabled.
+- Added `applyParentMessage` in `BaseMessage`.
+- Improved stability.
+
 ## v3.1.6 (DEC 29, 2021)
 - Bug-fix in retrying reconnect with application settings.
 - Fixed type definitions to `enum`.